Fix adjustDate to use long for microseconds (#415)

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/AbstractDateWithDay.java
@@ -79,7 +79,7 @@ abstract class AbstractDateWithDay extends AbstractDateWithMonth
     @Override
     public abstract AbstractDateWithDay clone();
 
-    void incrementDay(int delta)
+    void incrementDay(long delta)
     {
         if (delta < 0)
         {
